+++ src/main/java/net/brdle/collectorsreap/common/CRParticleTypes.java
@@ -13,6 +13,7 @@ public class CRParticleTypes {
 
 	public static final RegistryObject<SimpleParticleType> ACID = PARTICLE_TYPES.register("acid", () -> new SimpleParticleType(true));
 	public static final RegistryObject<SimpleParticleType> SHOCKWAVE = PARTICLE_TYPES.register("shockwave", () -> new SimpleParticleType(true));
+	public static final RegistryObject<SimpleParticleType> SURGE = PARTICLE_TYPES.register("surge", () -> new SimpleParticleType(true));
 
 	public static void create(IEventBus bus) {
 		PARTICLE_TYPES.register(bus);