+++ src/main/java/net/brdle/collectorsreap/common/effect/CREffects.java
@@ -1,8 +1,6 @@
 package net.brdle.collectorsreap.common.effect;
 
 import net.brdle.collectorsreap.CollectorsReap;
-import net.minecraft.core.particles.ParticleTypes;
-import net.minecraft.core.particles.SimpleParticleType;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.registries.DeferredRegister;
@@ -10,15 +8,16 @@ import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 
 public class CREffects {
-	public static final int MAX_SURGE = 4;
-	public static final float SURGE_CHANCE = 0.4F;
-	public static final SimpleParticleType SURGE_PARTICLE = ParticleTypes.ELECTRIC_SPARK;
+	public static final int MAX_SURGE = 2;
 
 	public static final DeferredRegister<MobEffect> EFFECTS = DeferredRegister.create(ForgeRegistries.MOB_EFFECTS, CollectorsReap.MODID);
 
-	public static final RegistryObject<MobEffect> CORROSION = EFFECTS.register("corrosion", CorrosionEffect::new);
-	public static final RegistryObject<MobEffect> VOLATILITY = EFFECTS.register("volatility", VolatilityEffect::new);
-	public static final RegistryObject<MobEffect> SURGE = EFFECTS.register("surge", SurgeEffect::new);
+	public static final RegistryObject<MobEffect> CORROSION = EFFECTS.register("corrosion",
+		CorrosionEffect::new);
+	public static final RegistryObject<MobEffect> VOLATILITY = EFFECTS.register("volatility",
+		VolatilityEffect::new);
+	public static final RegistryObject<MobEffect> SURGE = EFFECTS.register("surge",
+		SurgeEffect::new);
 
 	public static void create(IEventBus bus) {
 		EFFECTS.register(bus);