+++ src/main/java/net/brdle/collectorsreap/common/effect/SurgeEffect.java
@@ -1,5 +1,6 @@
 package net.brdle.collectorsreap.common.effect;
 
+import net.brdle.collectorsreap.common.CRParticleTypes;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.effect.MobEffect;
 import net.minecraft.world.effect.MobEffectCategory;
@@ -7,16 +8,12 @@ import net.minecraft.world.entity.LivingEntity;
 import org.jetbrains.annotations.NotNull;
 
 public class SurgeEffect extends MobEffect {
+
 	/**
-	 * Striking an enemy has a chance of summoning a lightning bolt
+	 * Striking an enemy can summon a lightning bolt
 	 */
 	protected SurgeEffect() {
 		super(MobEffectCategory.BENEFICIAL, 0xF1A3D2);
-	}
-
-	@Override
-	public void applyEffectTick(@NotNull LivingEntity entity, int amplifier) {
-		SurgeEffect.emitParticles(entity, amplifier);
 	}
 
 	/**
@@ -27,19 +24,26 @@ public class SurgeEffect extends MobEffect {
 		return duration % 10 == 0;
 	}
 
-	public static void emitParticles(@NotNull LivingEntity entity, int amplifier) {
+	@Override
+	public void applyEffectTick(@NotNull LivingEntity entity, int amplifier) {
+		if (amplifier == CREffects.MAX_SURGE) {
+			SurgeEffect.emitParticles(entity, 10);
+		}
+	}
+
+	public static void emitParticles(@NotNull LivingEntity entity, int amount) {
 		if (entity.level() instanceof ServerLevel server) {
-			for (int i = 0; i <= amplifier; i++) {
+			for (int i = 0; i < amount; i ++) {
 				server.sendParticles(
-					CREffects.SURGE_PARTICLE,
-					entity.getRandomX(0.6D),
-					entity.getRandomY() + 0.5D,
-					entity.getRandomZ(0.6D),
-					2,
-					0.1D,
-					0.1D,
-					0.1D,
-					0.15D
+					CRParticleTypes.SURGE.get(),
+					entity.getRandomX(0.5D),
+					entity.getRandomY() + 0.4D,
+					entity.getRandomZ(0.5D),
+					1,
+					0D,
+					0D,
+					0D,
+					0D
 				);
 			}
 		}