+++ changelog.md
@@ -16,7 +16,13 @@
 - Added Rosé Moon (for [Brewin' and Chewin'](https://www.curseforge.com/minecraft/mc-mods/brewin-and-chewin))
 - Added Reanimator's Garden (for [Brewin' and Chewin'](https://www.curseforge.com/minecraft/mc-mods/brewin-and-chewin) and [Atmospheric](https://www.curseforge.com/minecraft/mc-mods/atmospheric))
 - Added Pink Dragon Fruit Gummy
+- Lime Gummy now provides Corrosion for 0:10
+- Pomegranate Gummy now provides Volatility for 0:10
+- Apple Gummy now provides Comfort for 0:45
+- Vanilla Gummy now provides Vanilla Scent for 0:10
 - Glow Berry Gummy will now provide Tracer III in presence of [Cosmopolitan](https://www.curseforge.com/minecraft/mc-mods/cosmopolitan)
+- Urchin Dart can be crafted using `forge:nuggets/copper` if the tag is populated
+- Powder Cannon from [My Nether's Delight](https://www.curseforge.com/minecraft/mc-mods/my-nethers-delight) can be used as an Urchin Dart shooter
 - Added more Coaster item models for [Brewin' and Chewin'](https://www.curseforge.com/minecraft/mc-mods/brewin-and-chewin)
 
 ### Fixes